requirements.txt
python-dotenv~=1.0.0
python-keycloak~=3.7.0
mailjet-rest~=1.3.4